Added the following functionality:
when deleting an asset tag for an asset, delete it by tag name or by tag value.
be able to create an asset tag 'on the fly' without specifying a specific tag name (tag name defaults to 'GEN').
be able to find an asset by an asset tag value or by an asset tag name/value pair.
modified dimSim.jsp so it can do all of the above DAM functions.
added the path attribute to the asset tag in the XML.
added the GEN asset attribute in the XML with comma delimited values.
added the protect Asset/Folder Actions, making the files in the O/S read-only.

Fixed the following bugs:

When creating an asset, don't create the TITLE tag by default.
Since TITLE tag is no longer created by default then the updateAssetTitle function is deleted.
added the protect Asset/Folde functions in the Controller, which weren't there before.