Added the following functionality:
Virtual folders are created using both clientid and valveid as search criteria.
Added damSim simulations for protect/unprotect folder/asset and move asset to a folder.

Fixed the following bugs:

Remove all calls to getFolder by string name, changed to folder id.
A new issue that spin-offs from this task is the fact that is not safe to delete/move/changeTo/protect/unprotect a folder by it's name.
If the folder exists, must have an id and all of the above functions have been changed to be called with the folder id to avoid any confusion.
Since move asset involves a folder, must change this function to use folder id as oppose to folder name for the same reason specified above.

Modified damSim.jsp, AssetManagerTest and AssetManagerControllerTest to test the above changes.