@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ When this code revision gets moved to stagging, all assets, folders and tags in the DB must be deleted (Start from Scratch if there are any shared assets in the db) @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
Added the following functionality:

French compatablitlity for asset manager. Tags, titles, descriptions, folder names and files are allowed to be saved and retrieved and searched upon in French.

Fixed the following bugs:

When the asset is uploaded/moved to the ROOT directory the original valveid was still showing instead of "*" which is the equivalent to ALL_VALVES (all assets at the ROOT level are shared among valves).