@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ When this code revision gets moved to stagging, all assets, folders and tags in the DB must be deleted (Start from Scratch if there are any logos already created in the APEC folder) @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
Added the following functionality:

Created a system folder called "BZ Logo" that points to an APEC folder in the O/S. This folder contains a file called "logo.swf" which is the BZ created logo for the client. Upload the file as a read-only asset called "logo.swf" under the "BZ Logo" folder.

Fixed the following bugs:

When doing a findAssetsByName or findAssetsByTag, if the value searched has a single quote (as in French word l'arrière-boutique or as in English word Boutique's) it generates an exception.  Had to quote the quote in the HQL query.