Added the following functionality:
Return only assets that have the "manufacturer" prefix ignore all the other assets/folders.
Also, add the "OEM Logos" in the context, so that once it's been created, it will be saved in cache and reused and not re-created every time.

Fixed the following bugs:

get rid of the extra slash in protocol.